Homologar el resto de los tests en las nuevas carpetas
